New update

+++ Exercicios em C/Exercicio 7 - Dobro e Terça parte/A7.c
@@ -2,7 +2,7 @@
 
 float n1, dobro, tercaparte = 0;
 
-float main(){
+int main(){
 	
 	printf("Digite um numero: ");
 		scanf("%f", &n1);
@@ -10,6 +10,6 @@ float main(){
 	dobro = n1 * 2;
 	tercaparte = n1 / 3;
 	
-	printf("O dobro de %f e igual a %f\n ", n1, dobro);
-	printf("A terca parte de %f e igual a %f", n1, tercaparte);
+	printf("O dobro de %.1f e igual a %.2f\n ", n1, dobro);
+	printf("A terca parte de %.1f e igual a %.2f", n1, tercaparte);
 }
